refactor: move lib functions from store file (player.ts) to dedicated one
