+++ src/components/FavoritesView/FavoritesList.svelte
@@ -6,7 +6,6 @@
     playNextList,
     currentID,
     favoritesPlayStatus,
-    shuffle,
   } from '$lib/player';
   import { fade } from 'svelte/transition';
 
@@ -16,6 +15,7 @@
   import FavoritesItem from './FavoritesItem.svelte';
   import Modal from '$lib/Modal.svelte';
   import Orderable from '$lib/Orderable.svelte';
+  import { shuffle } from '$lib/shuffle';
 
   let playWarning = false;
   let warningAction = () => {};
