+++ src/lib/secondsToTime.ts
@@ -1,0 +1,8 @@
+export function secondsToTime(seconds: number) {
+    const h = Math.floor(seconds / 3600).toString().padStart(2, '0'),
+        m = Math.floor(seconds % 3600 / 60).toString().padStart(2, '0'),
+        s = Math.floor(seconds % 60).toString().padStart(2, '0');
+    if (h === '00')
+        return m + ':' + s;
+    return h + ':' + m + ':' + s;
+};
